requirements.txt
pandas 
numpy
scikit-learn
kagglehub
fastapi[all]
optuna
zenml
wandb
logfire
logfire[fastapi]
python-dotenv